Add simple video labelling apis. Small structure changes for acr test
