+++ listening-test-server/handlers/download_csv/video_labeling_csv_download.py
@@ -1,0 +1,7 @@
+from handlers.download_csv.ab_test_responses_download import AbTestResponsesDownload
+
+
+class VideoLabelingCsvDownload(AbTestResponsesDownload):
+    async def prepare(self):
+        self.user_id = await self.auth_current_user()
+        self.surveyCollectionName = 'videoLabelingSurveys'
