+++ listening-test-server/handlers/miscellanea/template_handler.py
@@ -1,6 +1,5 @@
 from typing import Optional
 
-import pymongo
 import tornado.web
 from pymongo.collection import Collection
 
@@ -54,6 +53,8 @@ def switch_test_collection(self: BaseHandler, test_type: str) -> Optional[Collec
         return self.db['audioLabelingTasks']
     elif test_type == 'image-labeling':
         return self.db['imageLabelingTasks']
+    elif test_type == 'video-labeling':
+        return self.db['videoLabelingTasks']
     else:
         self.set_error(400, 'Invalid task url')
         raise tornado.web.Finish
