+++ listening-test-server/handlers/miscellanea/test_responses.py
@@ -30,7 +30,7 @@ class TestResponsesHandler(BaseHandler):
             data = collection.find_one({'userId': self.user_id, '_id': ObjectId(_id)})
         # Check if the responses are downloadable
         if downloadable:
-            json_file_name = f"{self.test_name}_Test-{datetime.now().strftime('%Y%m%d%H%M%S')}.json"
+            json_file_name = f"{test_type}-{datetime.now().strftime('%Y%m%d%H%M%S')}.json"
             # Set http response header for downloading file
             self.set_header('Content-Type', 'application/octet-stream')
             self.set_header('Content-Disposition', f'attachment; filename="{json_file_name}"')
@@ -62,6 +62,8 @@ def switch_response_collection(self: BaseHandler, test_type: str) -> Optional[Co
         return self.db['audioLabelingSurveys']
     elif test_type == 'image-labeling':
         return self.db['imageLabelingSurveys']
+    elif test_type == 'video-labeling':
+        return self.db['videoLabelingSurveys']
     else:
         self.set_error(400, 'Invalid task url')
         raise tornado.web.Finish
