+++ listening-test-server/url.py
@@ -5,23 +5,24 @@ from handlers.download_csv.audio_labeling_csv_download import AudioLabelingCsvDo
 from handlers.download_csv.hearing_test_csv_download import HearingTestCsvDownload
 from handlers.download_csv.image_labeling_csv_download import ImageLabelingCsvDownload
 from handlers.download_csv.mushra_test_csv_download import MushraTestCsvDownload
+from handlers.download_csv.video_labeling_csv_download import VideoLabelingCsvDownload
 from handlers.miscellanea.storage_status import StorageStatusHandler
-from handlers.test_and_survey.acr_test import AcrTestHandler
+from handlers.test_and_survey.audio_acr_test import AcrTestHandler, AcrSurveyHandler
 from handlers.auth.sign_up_handler import SignUpHandler
 from handlers.auth.login import LoginHandler
 from handlers.auth.password import PasswordHandler
-from handlers.test_and_survey.ab_test import AbTestHandler, AbTestSurveyHandler
+from handlers.test_and_survey.audio_ab_test import AbTestHandler, AbTestSurveyHandler
 from handlers.download_csv.acr_test_csv_download import AcrTestCsvDownload
-from handlers.test_and_survey.acr_survey import AcrSurveyHandler
 from handlers.test_and_survey.audio_labeling_task import AudioLabelingHandler, AudioLabelingSurveyHandler
 from handlers.test_and_survey.hearing_test import HearingTestHandler, HearingSurveyHandler
 from handlers.test_and_survey.image_labeling_task import ImageLabelingHandler, ImageLabelingSurveyHandler
-from handlers.test_and_survey.mushra_test import MushraTestHandler, MushraSurveyHandler
+from handlers.test_and_survey.audio_mushra_test import MushraTestHandler, MushraSurveyHandler
 from handlers.miscellanea.responses_count import ResponsesCountHandler
 from handlers.download_csv.ab_test_responses_download import AbTestResponsesDownload
 from handlers.miscellanea.test_responses import TestResponsesHandler
 from handlers.file_handler import FileHandler
 from handlers.miscellanea.template_handler import TemplateHandler
+from handlers.test_and_survey.video_labeling_task import VideoLabelingHandler, VideoLabelingSurveyHandler
 
 path = [
     # Public
@@ -60,6 +61,9 @@ path = [
     ("/api/task/image-labeling", ImageLabelingSurveyHandler),
     ("/api/csv-download/image-labeling", ImageLabelingCsvDownload),
     # Video
+    ("/api/video-labeling", VideoLabelingHandler),
+    ("/api/task/video-labeling", VideoLabelingSurveyHandler),
+    ("/api/csv-download/video-labeling", VideoLabelingCsvDownload),
     # Response and Misc
     ("/api/response", TestResponsesHandler),
     ("/api/response-count", ResponsesCountHandler),
